better wasm split
